agent: Move models to submodule
